Now the code builds with -Wall -Werror -pedantic (because @eddelbuettel likes it that way)

+++ include/cpuaff/impl/basic_affinity_manager.hpp
@@ -152,7 +152,7 @@ class basic_affinity_manager
     {
         if (has_cpus())
         {
-            if (i >= 0 && i < cpu_by_index_.size())
+            if (i >= 0 && uint32_t(i) < cpu_by_index_.size())
             {
                 cpu = cpu_by_index_[i];
                 return true;
